klient/machine: move Change type to separate file
